# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API REST Panel de Control de Club Deportivo",
        "description": "API para autenticación y gestión de un club deportivo",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/v1/courts": {
            "get": {
                "tags": [
                    "Pistas"
                ],
                "summary": "Listar todas las pistas",
                "operationId": "dd5e9c4eb400bbef800707aebf805801",
                "responses": {
                    "200": {
                        "description": "Lista de pistas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CourtResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pistas"
                ],
                "summary": "Registrar una nueva pista",
                "operationId": "5ad21879b0592303f39a12078a74160e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCourtRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pista registrada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CourtResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/courts/{id}": {
            "get": {
                "tags": [
                    "Pistas"
                ],
                "summary": "Obtener una pista por ID",
                "operationId": "24d67680b98625e4a46640b06fbca301",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la pista",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos de la pista",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CourtResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pista no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Pistas"
                ],
                "summary": "Actualizar una pista",
                "operationId": "822dffd21225e33c274002ba4a74eee9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la pista",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCourtRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pista actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CourtResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pistas"
                ],
                "summary": "Eliminar una pista",
                "operationId": "a0d875024139da99678e6ebcfb4ded9d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la pista",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Pista eliminada correctamente"
                    },
                    "404": {
                        "description": "Pista no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Pistas"
                ],
                "summary": "Actualizar parcialmente una pista",
                "operationId": "5a51bc0c9024b5efd8da6e55c2d9eda2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la pista",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCourtRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pista actualizada parcialmente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CourtResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pista no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/courts/available": {
            "get": {
                "tags": [
                    "Pistas"
                ],
                "summary": "Buscar pistas disponibles",
                "operationId": "032c3429306609b1b3a589418ef9119b",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Fecha de la reserva en formato Y-m-d",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-02-28"
                        }
                    },
                    {
                        "name": "sport_id",
                        "in": "query",
                        "description": "ID del deporte para filtrar las pistas",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "member_id",
                        "in": "query",
                        "description": "ID del socio que realiza la consulta",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    },
                    {
                        "name": "time",
                        "in": "query",
                        "description": "Hora específica para verificar disponibilidad (opcional)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "time",
                            "example": "10:00"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de pistas disponibles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CourtResource"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Parámetros inválidos"
                    },
                    "500": {
                        "description": "Error en el servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreCourtRequest": {
                "title": "Store Court Request",
                "description": "Datos necesarios para crear una pista",
                "required": [
                    "name",
                    "sport_id"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Cancha 1"
                    },
                    "sport_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "UpdateCourtRequest": {
                "title": "Update Court Request",
                "description": "Datos para actualizar una pista",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Cancha 1"
                    },
                    "sport_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "CourtResource": {
                "title": "Court Resource",
                "description": "Datos de la pista deportiva",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Cancha 1"
                    },
                    "sport": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Tenis"
                            }
                        },
                        "type": "object"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-26T12:05:09Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-26T12:05:09Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}